수정/삭제 기능 수정 - 삭제 로직(handleTodoDelete) 수정 : TodoList title 동일한 경우 같이 삭제되는 문제 발생하여 filter메서드 내부 title 비교 삭제, 삭제 후 모달 창 닫히게 하기 위해 기존에는 state값 변경하였는데 모달 창 닫는 함수를 내부에서 호출하는 것으로 변경 - 수정 로직(handleTodoEdit) 수정 : 조건문으로 input value값 0인 경우 처리 - 상세페이지 크기 수정 : 상세페이지 크기 width,height px로 들어가있던 것 %로 수정 - 상세페이지 긴 문자 처리 : 상세페이지 input 텍스트 text-overflow: ellipsis 추가 - 커스텀 훅으로 분리 : useTodoList 내부로 수정/삭제 로직 이동 - 상세페이지 모달 오픈 시 애니메이션 추가
