+++ src/components/Detail/Detail.module.scss
@@ -4,34 +4,64 @@
 @use '/src/styles/mixins/position';
 
 .detailWrap {
-  position: absolute;
+  @include position.middleBox;
+  z-index: levels.$ADD_MODAL;
   width: 360px;
   height: 720px;
-  padding: 240px 40px 40px;
-  background-color: #fafbfe;
+  overflow: hidden;
   border-radius: 60px;
-  top: 0;
-  z-index: 11;
+
+  @keyframes fadeInDown {
+    0% {
+      transform: translateY(-30px);
+      opacity: 0;
+    }
+    to {
+      transform: translateY(0px);
+      opacity: 1;
+    }
+  }
+
+  .detail {
+    z-index: levels.$ADD_MODAL;
+    width: 100%;
+    height: 100%;
+    padding: 120px 50px 44px;
+    box-shadow: 0 0px 24px 8px rgba(colors.$WHITE, 95%);
+
+    & > * {
+      opacity: 0;
+      animation: fadeInDown 0.6s forwards;
+    }
+
+    @for $i from 1 through 5 {
+      & > *:nth-child(#{$i}) {
+        -webkit-animation-delay: ($i * 0.1) + 0.2s;
+        animation-delay: ($i * 0.1) + 0.2s;
+      }
+    }
+  }
 
   svg {
     width: 25px;
     height: 25px;
-    opacity: 0.2;
     cursor: pointer;
   }
 
   .title {
-    margin-bottom: 30px;
-    caret-color: colors.$BLUE;
+    width: 100%;
+    margin: 112px 0 48px 0;
     font-size: 30px;
+    text-overflow: ellipsis;
+    caret-color: colors.$BLUE;
   }
 
   .closeButton {
     @include position.absolute(40px 30px auto auto);
     width: 45px;
     height: 45px;
-    border-radius: 50%;
     border: 2px solid colors.$GRAYD;
+    border-radius: 50%;
     transition: 0.2s;
 
     &:hover {
@@ -43,8 +73,8 @@
       @include position.middleBox;
       width: 18px;
       height: 2px;
-      content: '';
       background-color: colors.$GRAY3;
+      content: '';
     }
 
     &::after {
@@ -66,8 +96,8 @@
       height: 50px;
       padding: 0 20px;
       color: colors.$GRAYA;
-      border-radius: 30px;
       border: 2px solid colors.$GRAYE;
+      border-radius: 30px;
     }
 
     .recordButton {
@@ -75,15 +105,15 @@
       width: 45px;
       height: 45px;
       margin-left: 10px;
-      border-radius: 50%;
       border: 2px solid colors.$GRAYE;
+      border-radius: 50%;
 
       &::before,
       &::after {
         @include position.posCenter;
         display: block;
-        content: '';
         border-radius: 50%;
+        content: '';
       }
 
       &::after {
@@ -107,6 +137,7 @@
 
     svg {
       margin: 0 20px;
+      opacity: 0.4;
     }
   }
 }
