+++ src/components/Detail/DetailBubble.js
@@ -1,0 +1,7 @@
+import styles from './DetailBubble.module.scss'
+
+function ExpandBubble() {
+  return <div className={styles.expandBubble} />
+}
+
+export default ExpandBubble
